Added defines for <stdint.h> and <stdbool.h>.
This adds HAVE_STDINT_H and HAVE_STDBOOL_H to the configuration.  We've
been using HAVE_STDINT_H for a while but just never bothered to define
it, and we're about to start using HAVE_STDBOOL_H.

I enabled them for the linux and darwin targets, and left it undefined
for freebsd/windows.