Make adb's daemon-port on the host machine configurable.
This is the first CL of a somewhat larger effort which, among other things,
will involve changing the emulator and ddms to talk to adb running on a
configurable port.

The port can be configured using environment variable ANDROID_ADB_PORT.

Further CLs will also address the set of ports used for the local transport.

Change-Id: I8082ee10ce504dab1cf89fe6d189eccf43ce9cf4