crash_collector: Handle compressed i915_error_state for chrome crashes
i915_error_state was previously truncated by debugd,
and will be compressed in the future to prevent that.

Because of compression, it will now be received in base64 format,
so to reduce it's size, decode it back.

Also, use xz compression, since it's more compact.
The second compression is a preparation for adding task states to the logs.

BUG=chromium:219121
TEST=Kill chrome, see the log gathered. Uploaded report id 5e5a6f9bf6ff4b5f.

Change-Id: I316cdaef8f8f42d4ac839f7c880960cdf6d7a1e8
Reviewed-on: https://gerrit.chromium.org/gerrit/62568
Reviewed-by: Albert Chaulk <achaulk@chromium.org>
Tested-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Yuly Novikov <ynovikov@chromium.org>