Add a config for dual mode zygote.
Note that init.zygote64.rc, which supports a "pure"
64 bit zygote is around only for testing.

The life cycles of both zygotes are controlled by init,
and the assumption here is that they will be available
always. We start the system_server in 32 bit mode.

Note that the distinction between "primary" and "secondary"
simply defines the order in which ABI support is queried,
there's no real requirement that the primary zygote supports
the primary ABI of the device.

bug: 13647418

Change-Id: Id0be001ea6f934c3c2022d89a63aae9fae66cc38